@@ -2,168 +2,183 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-using System.Net;
 using System.Threading;
-using Caliburn.Micro;
-using DayZ2.DayZ2Launcher.App.Ui;
-using DayZ2.DayZ2Launcher.App.Ui.Controls;
+using System.Threading.Tasks;
 
 namespace DayZ2.DayZ2Launcher.App.Core
 {
-	public class ServerList : ViewModelBase,
-		IHandle<RefreshServerRequest>
+	public class Filters
 	{
-		private bool _downloadingServerList;
-		private bool _isRunningRefreshBatch;
-		private ObservableCollection<Server> _items;
-		private ObservableCollection<Server> _oldItems;
 
-		private ServerBatchRefresher _refreshAllBatch;
+	}
 
-		public ServerList()
+	public struct Player
+	{
+		public string Name;
+		public int Score;
+		public string Duration;
+	}
+
+	public enum ServerDifficulty
+	{
+		Recruit = 0,
+		Regular = 1,
+		Veteran = 2,
+		Mercenary = 3
+	}
+
+	public enum ServerState
+	{
+		None,
+		SelectingMission,
+		EditingMission,
+		AssigningRoles,
+		SendingMission,
+		LoadingGame,
+		Briefing,
+		Playing,
+		Debriefing,
+		MissionAborted
+	}
+
+	public enum ServerPlatform
+	{
+		Undefined = 0,
+		Windows = 'w',
+		Linux = 'l'
+	}
+
+	public enum ServerPerspective
+	{
+		FirstPerson,
+		ThirdPerson
+	}
+
+	public class Server
+	{
+		public string Name;
+		public readonly string Hostname;
+		public readonly ushort GamePort;
+		public readonly ushort QueryPort;
+		public bool IsResponding;
+		public string Version;
+		public string RequiredVersion;
+		public string RequiredBuildNo;
+		public bool EqualModRequired;
+		public bool Lock;
+		public bool VerifySignatures;
+		public bool Battleye = false;
+		public bool Dedicated;
+		public long? Ping;
+		public int Slots;
+		public int PlayerCount;
+		public int FreeSlots => Slots - PlayerCount;
+		public IList<Player> Players;
+		public IList<string> Mods;
+		public ServerDifficulty Difficulty;
+		public ServerPerspective Perspective;
+		public ServerState State;
+		public string GameType;
+		public string Language;
+		public string LatLong;
+		public ServerPlatform Platform = ServerPlatform.Undefined;
+		public string ContentLoadedHash;
+		public string Country;
+		public string TimeLeft;
+		public bool IsFavorite;
+
+		public event EventHandler<EventArgs> RefreshStarted;
+		public event EventHandler<EventArgs> RefreshFinished;
+		public event EventHandler<EventArgs> Refreshed;
+
+		public Server(ServerListInfo info)
 		{
-			Items = new ObservableCollection<Server>();
+			Hostname = info.Hostname;
+			GamePort = info.Port;
+			QueryPort = (ushort)(info.Port + 1);
+			Mods = info.Mods;
 		}
 
-		public ServerBatchRefresher RefreshAllBatch
+		public async Task RefreshAsync(CancellationToken cancellationToken)
 		{
-			get { return _refreshAllBatch; }
-			private set
+			try
 			{
-				_refreshAllBatch = value;
-				PropertyHasChanged("RefreshAllBatch");
+				RefreshStarted?.Invoke(this, EventArgs.Empty);
+				await ArmaServerQuery.Run(this, cancellationToken);
+				IsResponding = true;
 			}
-		}
-
-		public ObservableCollection<Server> Items
-		{
-			get { return _items; }
-			private set
+			catch (TimeoutException)
 			{
-				_items = value;
-				PropertyHasChanged("Items");
+				IsResponding = false;
 			}
-		}
-
-		public bool DownloadingServerList
-		{
-			get { return _downloadingServerList; }
-			set
+			finally
 			{
-				_downloadingServerList = value;
-				PropertyHasChanged("DownloadingServerList");
+				RefreshFinished?.Invoke(this, EventArgs.Empty);
 			}
-		}
-
-		public void Handle(RefreshServerRequest message)
-		{
-			if (_isRunningRefreshBatch)
-				return;
-
-			_isRunningRefreshBatch = true;
-			App.Events.Publish(new RefreshingServersChange(true));
-			RefreshAllBatch = message.Batch;
-			RefreshAllBatch.RefreshAllComplete += RefreshAllBatchOnRefreshAllComplete;
-			RefreshAllBatch.RefreshAll();
-		}
-
-		public void GetAndUpdateAll()
-		{
-			GetAll(() => UpdateAll());
-		}
-
-		private void GetAll(Action uiThreadOnComplete)
-		{
-			DownloadingServerList = true;
-			new Thread(() =>
-			{
-				List<Server> servers = GetAllSync();
-				Execute.OnUiThread(() =>
-				{
-					_oldItems = Items;
-					Items = new ObservableCollection<Server>(servers);
-					DownloadingServerList = false;
-					if (uiThreadOnComplete != null)
-						uiThreadOnComplete();
-				});
-			}).Start();
-		}
-
-		private List<Server> GetAllSync()
-		{
-			string list = "";
-			{
-				string serverListUrl = "https://www.perry-swift.de/dayz2/serverlist.txt";
-				// LocatorInfo locator = CalculatedGameSettings.Current.Locator;
-				//if (locator != null && locator.ServerListUrl != null)
-				//	serverListUrl = locator.ServerListUrl;
-
-				if (!string.IsNullOrWhiteSpace(serverListUrl))
-				{
-					using (var wc = new WebClient())
-					{
-						try
-						{
-							list = wc.DownloadString(new Uri(serverListUrl));
-						}
-						catch (Exception)
-						{
-						}
-					}
-				}
-			}
-
-			if (string.IsNullOrEmpty(list))
-				return new List<Server>(); //Empty list.. Too bad.
-
-			List<Server> fullList = list
-				.Split('\n').Select(line =>
-				{
-					string[] serverInfo = line.Split(';');
-					var server = new Server("", 0, "", "", 0);
-					if (serverInfo.Count() > 5)
-					{
-						string queryHostname = serverInfo[1];
-						ushort joinPort = (ushort)serverInfo[2].TryInt();
-						string password = serverInfo[3];
-						string mod = serverInfo[4];
-						ushort queryPort = (ushort)serverInfo[5].TryInt();
-
-						server = new Server(queryHostname, joinPort, password, mod, queryPort);
-					}
-
-					server.Settings = new SortedDictionary<string, string>
-					{
-						{"hostname", serverInfo[0]}
-					};
-
-					return server;
-				}).ToList();
-
-			return fullList;
-		}
-
-		public void UpdateAll()
-		{
-			var batch = new ServerBatchRefresher("Refreshing all servers...", Items, _oldItems);
-			App.Events.Publish(new RefreshServerRequest(batch));
-		}
-
-		private void RefreshAllBatchOnRefreshAllComplete()
-		{
-			RefreshAllBatch.RefreshAllComplete -= RefreshAllBatchOnRefreshAllComplete;
-			_isRunningRefreshBatch = false;
-			App.Events.Publish(new RefreshingServersChange(false));
+			Refreshed?.Invoke(this, EventArgs.Empty);
 		}
 	}
 
-	public class RefreshingServersChange
+
+	public class ServerDiscoveredEventArgs : EventArgs
 	{
-		public RefreshingServersChange(bool isRunning)
+		public Server Server { get; private set; }
+
+		public ServerDiscoveredEventArgs(Server server)
 		{
-			IsRunning = isRunning;
+			Server = server;
+		}
+	}
+
+	public class ServerList
+	{
+		public List<Server> Servers { get; private set; } = new();
+
+		//public IList<Server> Servers { get; private set; }
+		public IList<Server> FilteredServers { get; private set; }
+
+		public bool IsRefreshing { get; private set; }
+
+		public event EventHandler<ServerDiscoveredEventArgs> ServerDiscovered;
+
+		public void Join(int index)
+		{
+
 		}
 
-		public bool IsRunning { get; set; }
+		public void SetServers(IList<ServerListInfo> servers)
+		{
+			Servers.Clear();
+			foreach (ServerListInfo info in servers)
+			{
+				Server server = new Server(info);
+				Servers.Add(server);
+				ServerDiscovered?.Invoke(this, new ServerDiscoveredEventArgs(server));
+			}
+		}
+
+		public void ApplyFilter(Func<Server, bool> filter)
+		{
+			FilteredServers = Servers.Where(filter).ToArray();
+		}
+
+		public void ResetFilters()
+		{
+			FilteredServers = Servers;
+		}
+
+		public async Task RefreshAsync(int index, CancellationToken cancellationToken)
+		{
+			IsRefreshing = true;
+			await Servers[index].RefreshAsync(cancellationToken);
+			IsRefreshing = false;
+		}
+
+		public async Task RefreshAllAsync(CancellationToken cancellationToken)
+		{
+			IsRefreshing = true;
+			await Task.WhenAll(Servers.Select(s => s.RefreshAsync(cancellationToken)));
+			IsRefreshing = false;
+		}
 	}
 }
