@@ -1,28 +1,27 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Data;
 
 namespace DayZ2.DayZ2Launcher.App.Ui.Converters
 {
-	public class DifficultyToDescriptionConverter : IValueConverter
+	class BooleanToVisibilityConverter : IValueConverter
 	{
 		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
 		{
-			string nullText = System.Convert.ToString(parameter);
+			if (value != null)
+			{
+				if ((bool)value)
+					return Visibility.Visible;
+				else
+					return Visibility.Hidden;
+			}
 
-			var val = (int?)value;
-			if (val == null)
-				return nullText;
-			if (val == 0)
-				return "Recruit";
-			if (val == 1)
-				return "Regular";
-			if (val == 2)
-				return "Veteran";
-			if (val == 3)
-				return "Expert";
-
-			return nullText;
+			return Visibility.Hidden;
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
