@@ -7,11 +7,12 @@ namespace DayZ2.DayZ2Launcher.App.Ui.Converters
 {
 	public class PingToForegroundConverter : IValueConverter
 	{
+		private static SolidColorBrush Nothing = new SolidColorBrush(Color.FromArgb(255, 204, 204, 204));
 		private static SolidColorBrush Fastest = new SolidColorBrush(Color.FromArgb(255, 25, 253, 25));
 		private static SolidColorBrush Fast = new SolidColorBrush(Color.FromArgb(255, 120, 239, 120));
 		private static SolidColorBrush Medium = new SolidColorBrush(Colors.Yellow);
-		private static SolidColorBrush Slow = new SolidColorBrush(Colors.Red);
-		private static SolidColorBrush Nothing = new SolidColorBrush(Color.FromArgb(255, 204, 204, 204));
+		private static SolidColorBrush Slow = new SolidColorBrush(Colors.Orange);
+		private static SolidColorBrush Slowest = new SolidColorBrush(Colors.Red);
 
 		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
 		{
@@ -19,14 +20,16 @@ namespace DayZ2.DayZ2Launcher.App.Ui.Converters
 				return Nothing;
 
 			var val = (long)value;
-			if (val > 0 && val < 100)
+			if (val < 60)
 				return Fastest;
-			if (val >= 100 && val < 160)
+			if (val < 120)
 				return Fast;
-			if (val >= 160 && val < 275)
+			if (val < 180)
 				return Medium;
+			if (val < 240)
+				return Slow;
 
-			return Slow;
+			return Slowest;
 		}
 
 		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
