@@ -1,6 +1,7 @@
 using System;
 using System.Collections.ObjectModel;
 using System.Linq;
+using System.Threading;
 using System.Threading.Tasks;
 using DayZ2.DayZ2Launcher.App.Core;
 using DayZ2.DayZ2Launcher.App.Ui.ServerList;
@@ -9,82 +10,72 @@ namespace DayZ2.DayZ2Launcher.App.Ui
 {
 	public class MainWindowViewModel : ViewModelBase
 	{
-		private ViewModelBase _currentTab;
-		private GameLauncher_old _launcher;
-		private Core.ServerList _serverList;
-		private ObservableCollection<ViewModelBase> _tabs;
+		private CancellationToken m_cancellationToken;
 
-		public MainWindowViewModel()
+		private ViewModelBase m_currentTab;
+		private ObservableCollection<ViewModelBase> m_tabs;
+
+		public MainWindowViewModel(IServiceProvider services, AppCancellation cancellation)
 		{
+			m_cancellationToken = cancellation.Token;
+
+			GameLauncher = new GameLauncher();
+
 			Tabs = new ObservableCollection<ViewModelBase>(new ViewModelBase[]
 			{
-				ServerListViewModel = new ServerListViewModel()
+				ServerListViewModel = services.CreateInstance<ServerListViewModel>(),
+				UpdatesViewModel = services.CreateInstance<UpdatesViewModel>(),
+				SettingsViewModel = services.CreateInstance<SettingsViewModel>(),
 			});
 			CurrentTab = Tabs.First();
 
-			ServerList = new Core.ServerList();
-			Launcher = new GameLauncher_old();
-
-			SettingsViewModel = new SettingsViewModel();
-			UpdatesViewModel = new UpdatesViewModel(Launcher);
+			// SettingsViewModel = new SettingsViewModel();
+			// UpdatesViewModel = new UpdatesViewModel(GameLauncher, m_cancellationToken);
 
 			SettingsViewModel.TorrentSettingsChanged += (sender, args) =>
 			{
 				UpdatesViewModel.ReconfigureTorrentEngine();
 			};
 
-			ServerListViewModel.Launcher = Launcher;
-		}
-
-
-		public Core.ServerList ServerList
-		{
-			get => _serverList;
-			set
+			UpdatesViewModel.PropertyChanged += (sender, args) =>
 			{
-				_serverList = value;
-				PropertyHasChanged(nameof(ServerList));
-			}
-		}
-
-		public GameLauncher_old Launcher
-		{
-			get => _launcher;
-			set
-			{
-				_launcher = value;
-				PropertyHasChanged(nameof(Launcher));
-			}
+				if (args.PropertyName == nameof(UpdatesViewModel.Servers))
+				{
+					ServerListViewModel.SetServers(UpdatesViewModel.Servers);
+				}
+			};
 		}
 
 		public ServerListViewModel ServerListViewModel { get; set; }
 		public SettingsViewModel SettingsViewModel { get; set; }
 		public UpdatesViewModel UpdatesViewModel { get; set; }
+		public GameLauncher GameLauncher { get; set; }
 
 		public ViewModelBase CurrentTab
 		{
-			get { return _currentTab; }
+			get => m_currentTab;
 			set
 			{
-				if (_currentTab != null)
-					_currentTab.IsSelected = false;
-				_currentTab = value;
-				if (_currentTab != null)
-					_currentTab.IsSelected = true;
-				PropertyHasChanged(nameof(CurrentTab), nameof(IsServerListSelected));
+				if (m_currentTab != null)
+					m_currentTab.IsSelected = false;
+				m_currentTab = value;
+				if (m_currentTab != null)
+					m_currentTab.IsSelected = true;
+				OnPropertyChanged(nameof(CurrentTab), nameof(IsServerListSelected));
 			}
+		}
+
+		public void Shutdown()
+		{
+			//TODO: cancel in App
 		}
 
 		public bool IsServerListSelected => CurrentTab == ServerListViewModel;
 
 		public ObservableCollection<ViewModelBase> Tabs
 		{
-			get => _tabs;
-			set
-			{
-				_tabs = value;
-				PropertyHasChanged(nameof(Tabs));
-			}
+			get => m_tabs;
+			set => SetValue(ref m_tabs, value);
 		}
 
 		public void ShowSettings()
@@ -97,12 +88,6 @@ namespace DayZ2.DayZ2Launcher.App.Ui
 		{
 			SettingsViewModel.IsVisible = false;
 			UpdatesViewModel.IsVisible = true;
-		}
-
-		public void ShowPlugins()
-		{
-			SettingsViewModel.IsVisible = false;
-			UpdatesViewModel.IsVisible = false;
 		}
 
 		public void Escape()
